renamed qf builder
