+++ frontend/components/Domain/QueryFilterBuilder.vue
